working on authentication
