+++ Movie-Rater/client/src/pages/Home.jsx
@@ -1,15 +1,15 @@
-import { Link } from 'react-router-dom';
-import { useQuery } from'@apollo/client';
-import {QUERY_MOVIE } from '../utils/queries';
-import { React } from 'react';
+// import { Link } from 'react-router-dom';
+// import { useQuery } from'@apollo/client';
+// import {QUERY_MOVIE } from '../utils/queries';
+// import { React } from 'react';
 
 
 
-const Home = () => {
+// const Home = () => {
 
-    return (
-        <div></div>
-    )
-}
+//     return (
+//         <div></div>
+//     )
+// }
 
-export default Home;
+// export default Home;
