+++ Movie-Rater/vite.config.js
@@ -8,4 +8,4 @@ export default defineConfig({
     port: 3000,
     open: true
   }
-})
+});
